requirements.txt
torch
torchvision
numpy
matplotlib
natsort
Pillow
pandas
tqdm
opencv-python
